ヒートマップの修正
縦軸に系統，ドロップダウンメニューに都道府県名が入るよう修正しました．
PlotComponentのデータ処理をカスタムフックに移行し、レイアウト設定を整理しました。また、HeatmapPageでの系統表示数を増やし、データ取得ロジックを改善しました。docker-compose.ymlにソースコードの変更検知用ボリュームを追加し、next.config.jsでのホットリロード設定を強化しました。
