+++ src/app/visualization/components/PlotComponent.tsx
@@ -2,6 +2,8 @@
 
 import { useEffect, useState, useMemo } from 'react';
 import { Layout, Data } from 'plotly.js';
+import { usePlotDataProcess } from '@/hooks/usePlotDataprocess';
+import { usePlotLayout } from '@/hooks/usePlotLayout';
 
 interface PlotComponentProps {
   data: Data[];
@@ -12,125 +14,11 @@ export default function PlotComponent({ data, layout }: PlotComponentProps) {
   const [mounted, setMounted] = useState(false);
   const [Plot, setPlot] = useState<any>(null);
   const [showDebug, setShowDebug] = useState(false);
-  const [dataError, setDataError] = useState<string | null>(null);
 
-  // データの処理
-  const enhancedData = useMemo(() => {
-    try {
-      // データが空の場合は空配列を返す
-      if (!Array.isArray(data) || data.length === 0) {
-        return [];
-      }
+  const { enhancedData, dataError } = usePlotDataProcess(data);
+  const { enhancedLayout } = usePlotLayout(layout)
 
-      // 散布図（線グラフ）の場合
-      if (data.some((d: any) => d.type === 'scatter')) {
-        return data.map((trace: any) => {
-          // X軸とY軸のデータがあるか確認
-          const hasX = Array.isArray(trace.x) && trace.x.length > 0;
-          const hasY = Array.isArray(trace.y) && trace.y.length > 0;
-          
-          if (!hasX || !hasY) {
-            console.warn('散布図データにx/y配列がありません', trace);
-          }
-          
-          // 全ての系統名に対して同じスタイルを適用する
-          const colorForTrace = trace.line?.color || trace.marker?.color || undefined;
-          const lineWidth = 2;
-          const markerSize = 6;
-          
-          return {
-            ...trace,
-            visible: true,
-            showlegend: true,
-            opacity: 1,
-            line: {
-              ...(trace.line || {}),
-              width: lineWidth,
-              color: colorForTrace
-            },
-            marker: {
-              ...(trace.marker || {}),
-              size: markerSize,
-              opacity: 1,
-              color: colorForTrace
-            },
-            mode: trace.mode || 'lines+markers'
-          };
-        });
-      }
-      
-      // ヒートマップの場合
-      if (data.some((d: any) => d.type === 'heatmap')) {
-        return data.map((trace: any) => {
-          // z値（ヒートマップのデータ）があるか確認
-          const hasZ = Array.isArray(trace.z) && trace.z.length > 0;
-          
-          if (!hasZ) {
-            console.warn('ヒートマップデータにz配列がありません', trace);
-          }
-          
-          return {
-            ...trace,
-            visible: true,
-            showscale: true,
-            colorscale: trace.colorscale || 'Viridis'
-          };
-        });
-      }
-      
-      // その他の場合（もしくは特定できない場合）
-      return data.map((trace: any) => ({
-        ...trace,
-        visible: true,
-        showlegend: true
-      }));
-    } catch (error) {
-      console.error('データ処理エラー:', error);
-      return [];
-    }
-  }, [data]);
-
-  // エラー処理
-  useEffect(() => {
-    if (!Array.isArray(data) || data.length === 0) {
-      setDataError(null);
-      return;
-    }
-    
-    // データのバリデーション
-    try {
-      const hasScatter = data.some((d: any) => d.type === 'scatter');
-      const hasHeatmap = data.some((d: any) => d.type === 'heatmap');
-      
-      if (hasScatter) {
-        const invalidData = data.some((trace: any) => 
-          !Array.isArray(trace.x) || !Array.isArray(trace.y) || 
-          trace.x.length === 0 || trace.y.length === 0
-        );
-        
-        if (invalidData) {
-          setDataError('散布図データにx/y値が不足しています');
-          return;
-        }
-      }
-      
-      if (hasHeatmap) {
-        const invalidData = data.some((trace: any) => 
-          !Array.isArray(trace.z) || trace.z.length === 0
-        );
-        
-        if (invalidData) {
-          setDataError('ヒートマップデータにz値が不足しています');
-          return;
-        }
-      }
-      
-      setDataError(null);
-    } catch (error) {
-      console.error('データ検証エラー:', error);
-      setDataError('データの検証中にエラーが発生しました');
-    }
-  }, [data]);
+  
 
   useEffect(() => {
     console.log('===== PLOTLY DEBUG INFO =====');
@@ -148,132 +36,21 @@ export default function PlotComponent({ data, layout }: PlotComponentProps) {
 
   if (!mounted || !Plot) {
     return (
-      <div style={{
-        width: '100%',
-        height: '600px',
-        background: '#fff',
-        padding: '20px',
-        border: '1px solid #ddd',
-        borderRadius: '4px',
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center'
-      }}>
-        <p style={{ color: '#000000', fontWeight: '500' }}>グラフを読み込んでいます...</p>
+      <div className="w-full h-[600px] bg-white p-5 border border-gray-300 rounded-md flex justify-center items-center">
+        <p className="text-black font-medium">グラフを読み込んでいます...</p>
       </div>
     );
   }
 
-  // デフォルトのレイアウト設定とプロップスから渡されたレイアウトをマージ
-  const enhancedLayout = {
-    ...layout,
-    title: {
-      ...(layout.title || {}),
-      text: (layout.title as any)?.text || 'グラフ表示'
-    },
-    autosize: true,
-    width: 900, // 明示的な幅を設定
-    height: 500, // 明示的な高さを設定
-    paper_bgcolor: '#ffffff',
-    plot_bgcolor: '#ffffff',
-    showlegend: true,
-    font: {
-      ...layout.font,
-      color: '#000000',
-    },
-    legend: {
-      ...layout.legend,
-      font: {
-        ...(layout.legend?.font || {}),
-        color: '#000000',
-        size: 14
-      },
-      bgcolor: '#ffffff',
-      bordercolor: '#cccccc',
-      borderwidth: 1
-    },
-    modebar: {
-      bgcolor: '#ffffff',
-      color: '#000000',
-      activecolor: '#000000',
-    },
-    // データの範囲を強制的に拡大して見やすくする
-    xaxis: {
-      ...layout.xaxis,
-      autorange: true,
-      showgrid: true,
-      gridcolor: '#e0e0e0',
-      zeroline: true,
-      zerolinecolor: '#000000',
-      zerolinewidth: 1,
-      tickfont: {
-        ...(layout.xaxis?.tickfont || {}),
-        color: '#000000',
-        size: 12
-      },
-      title: {
-        ...(layout.xaxis?.title || {}),
-        font: {
-          ...(layout.xaxis?.title?.font || {}),
-          color: '#000000',
-          size: 14
-        }
-      }
-    },
-    yaxis: {
-      ...layout.yaxis,
-      autorange: true,
-      showgrid: true,
-      gridcolor: '#e0e0e0',
-      zeroline: true,
-      zerolinecolor: '#000000',
-      zerolinewidth: 1,
-      tickfont: {
-        ...(layout.yaxis?.tickfont || {}),
-        color: '#000000',
-        size: 12
-      },
-      title: {
-        ...(layout.yaxis?.title || {}),
-        font: {
-          ...(layout.yaxis?.title?.font || {}),
-          color: '#000000',
-          size: 14
-        }
-      }
-    },
-    annotations: layout.annotations?.map((annotation: any) => ({
-      ...annotation,
-      font: {
-        ...(annotation.font || {}),
-        color: '#000000'
-      }
-    })) || [],
-  };
+
 
   const isDataEmpty = !Array.isArray(enhancedData) || enhancedData.length === 0;
 
   return (
-    <div style={{
-      width: '100%',
-      height: '600px',
-      background: '#fff',
-      padding: '20px',
-      border: '1px solid #ddd',
-      borderRadius: '4px',
-      position: 'relative',
-      overflow: 'visible'
-    }}>
+    <div className="w-full h-[600px] bg-white p-5 border border-gray-300 rounded-md relative overflow-visible">
       {isDataEmpty ? (
-        <div style={{
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          height: '100%',
-          flexDirection: 'column',
-          gap: '10px'
-        }}>
-          <p style={{ color: '#000000', fontWeight: 'bold' }}>データがありません</p>
+        <div className="flex justify-center items-center h-full flex-col gap-2.5">
+          <p className="text-black font-bold">データがありません</p>
         </div>
       ) : (
         <Plot
@@ -291,81 +68,32 @@ export default function PlotComponent({ data, layout }: PlotComponentProps) {
           }}
         />
       )}
-      <div style={{
-        position: 'absolute',
-        top: '10px',
-        right: '10px',
-        display: 'flex',
-        flexDirection: 'column',
-        gap: '5px'
-      }}>
+      <div className="absolute top-2.5 right-2.5 flex flex-col gap-1.5">
         <button 
           onClick={() => setShowDebug(!showDebug)}
-          style={{
-            padding: '5px 10px',
-            background: '#f0f0f0',
-            border: '1px solid #000',
-            borderRadius: '4px',
-            cursor: 'pointer',
-            fontSize: '12px',
-            color: '#000000',
-            fontWeight: '500'
-          }}
+          className="px-2.5 py-1.5 bg-gray-100 border border-black rounded cursor-pointer text-xs text-black font-medium"
         >
           {showDebug ? 'デバッグ情報を隠す' : 'デバッグ情報を表示'}
         </button>
       </div>
       
       {dataError && (
-        <div style={{
-          position: 'absolute',
-          top: '50%',
-          left: '50%',
-          transform: 'translate(-50%, -50%)',
-          background: 'rgba(255,0,0,0.1)',
-          border: '1px solid #f00',
-          borderRadius: '4px',
-          padding: '10px',
-          color: '#f00'
-        }}>
+        <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 bg-red-100/10 border border-red-500 rounded p-2.5 text-red-500">
           {dataError}
         </div>
       )}
       
       {showDebug && (
-        <div style={{
-          position: 'absolute',
-          bottom: '50px',
-          right: '10px',
-          width: '300px',
-          maxHeight: '400px',
-          overflow: 'auto',
-          background: 'rgba(255,255,255,0.9)',
-          border: '1px solid #000',
-          padding: '10px',
-          fontSize: '10px',
-          whiteSpace: 'pre-wrap',
-          zIndex: 1000,
-          color: '#000000'
-        }}>
-          <h4 style={{ color: '#000000', fontWeight: 'bold', marginBottom: '8px' }}>データ情報:</h4>
-          <p style={{ color: '#000000', margin: '4px 0' }}>データセット数: {enhancedData.length}</p>
+        <div className="absolute bottom-[50px] right-2.5 w-[300px] max-h-[400px] overflow-auto bg-white/90 border border-black p-2.5 text-[10px] whitespace-pre-wrap z-[1000] text-black">
+          <h4 className="text-black font-bold mb-2">データ情報:</h4>
+          <p className="text-black my-1">データセット数: {enhancedData.length}</p>
           {!isDataEmpty && (
             <>
-              <p style={{ color: '#000000', margin: '4px 0' }}>データタイプ: {(enhancedData[0] as any)?.type || '不明'}</p>
-              <div style={{ marginTop: '10px' }}>
+              <p className="text-black my-1">データタイプ: {(enhancedData[0] as any)?.type || '不明'}</p>
+              <div className="mt-2.5">
                 <button
                   onClick={() => console.log('Debug Data:', enhancedData)}
-                  style={{
-                    padding: '3px 6px',
-                    fontSize: '9px',
-                    background: '#f5f5f5',
-                    border: '1px solid #000',
-                    borderRadius: '3px',
-                    cursor: 'pointer',
-                    color: '#000000',
-                    fontWeight: '500'
-                  }}
+                  className="py-0.5 px-1.5 text-[9px] bg-gray-100 border border-black rounded cursor-pointer text-black font-medium"
                 >
                   データをコンソールに出力
                 </button>
@@ -373,12 +101,12 @@ export default function PlotComponent({ data, layout }: PlotComponentProps) {
               
               {enhancedData.map((trace: any, i) => (
                 <div key={i}>
-                  <p style={{ color: '#000000', fontWeight: 'bold', marginTop: '8px' }}>データセット {i+1}: {trace.name || '名前なし'}</p>
-                  <p style={{ color: '#000000' }}>タイプ: {trace.type || '不明'}</p>
-                  {trace.x && <p style={{ color: '#000000' }}>x値サンプル: {Array.isArray(trace.x) ? trace.x.slice(0, 3).join(', ') + '...' : 'なし'}</p>}
-                  {trace.y && <p style={{ color: '#000000' }}>y値サンプル: {Array.isArray(trace.y) ? trace.y.slice(0, 3).join(', ') + '...' : 'なし'}</p>}
-                  {trace.z && <p style={{ color: '#000000' }}>z値（ヒートマップ）: あり</p>}
-                  <hr style={{ border: '0.5px solid #ccc', margin: '8px 0' }} />
+                  <p className="text-black font-bold mt-2">データセット {i+1}: {trace.name || '名前なし'}</p>
+                  <p className="text-black">タイプ: {trace.type || '不明'}</p>
+                  {trace.x && <p className="text-black">x値サンプル: {Array.isArray(trace.x) ? trace.x.slice(0, 3).join(', ') + '...' : 'なし'}</p>}
+                  {trace.y && <p className="text-black">y値サンプル: {Array.isArray(trace.y) ? trace.y.slice(0, 3).join(', ') + '...' : 'なし'}</p>}
+                  {trace.z && <p className="text-black">z値（ヒートマップ）: あり</p>}
+                  <hr className="border-t border-gray-300 my-2" />
                 </div>
               ))}
             </>
