+++ src/app/visualization/heatmap/[wave]/page.tsx
@@ -58,7 +58,7 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
   }
   
   // 出現頻度が高い系統のみを表示（パフォーマンス向上のため）
-  if (lineages.length > 5) {
+  if (lineages.length > 10) {
     // 各系統の出現回数をカウント
     const lineageCounts = lineages.map(lineage => {
       const count = data.filter(d => d.lineage.name === lineage).length;
@@ -67,7 +67,7 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
     
     // 出現回数で降順ソートして上位5のみを選択
     lineageCounts.sort((a, b) => b.count - a.count);
-    lineages = lineageCounts.slice(0, 5).map(item => item.lineage);
+    lineages = lineageCounts.slice(0, 10).map(item => item.lineage);
     console.log(`表示する系統を${lineages.length}個に制限しました`);
   }
 
@@ -87,38 +87,46 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
     );
   }
 
-  // 表示対象の都道府県を最大15に制限（パフォーマンス向上のため）
-  const topPrefectures = prefectures.length > 15 ? prefectures.slice(0, 15) : prefectures;
 
-  // 各系統ごとのヒートマップデータを作成
-  const heatmapData = lineages.map(lineage => {
-    const z = topPrefectures.map(prefecture => {
+
+  // 各系統（lineage）ごとにヒートマップデータを作成
+  // lineages配列内の各系統に対して処理を実行
+  const heatmapData = prefectures.map(prefecture => {
+    // z配列は2次元配列（ヒートマップの値を格納）
+    // 縦軸（y）は系統、横軸（x）は日付
+    const z = lineages.map(lineage => {
+      // 各都道府県に対して、すべての日付でのデータポイントを生成
       return dates.map(date => {
+        // 系統、都道府県、日付の組み合わせに一致するデータを検索
         const entry = data.find((d: CovidDataWithRelations) => 
           d.prefecture.name === prefecture && 
           d.lineage.name === lineage && 
           d.date.toISOString().split('T')[0] === date
         );
+        // データが存在する場合はその比率をパーセント表示に変換して返す、存在しない場合は0を返す
         return entry ? entry.ratio * 100 : 0;
       });
     });
 
+    // Plotly.jsに渡すヒートマップ設定オブジェクトを返す
     return {
-      z,
+      z, // 上で生成した2次元配列をセット
       x: dates.map(d => {
+        // 日付データを「年/週番号週」の形式に変換
         const date = new Date(d as string);
+        // 月初からの経過日数と月初の曜日から週番号を計算
         const weekNumber = Math.ceil((date.getDate() - 1 + new Date(date.getFullYear(), date.getMonth(), 1).getDay()) / 7);
         return `${date.getFullYear()}/${weekNumber}週`;
       }),
-      y: topPrefectures,
-      type: 'heatmap' as const,
-      colorscale: 'Viridis',
-      name: lineage,
-      visible: false,
-      showscale: true,
+      y: lineages, // y軸には都道府県名を設定
+      type: 'heatmap' as const, // グラフタイプはヒートマップ
+      colorscale: 'Viridis', // カラースケールはViridis（青紫〜黄緑のグラデーション）
+      name: prefectures, // 系統名を名前として設定
+      visible: false, // 初期状態では非表示（後で最初の要素のみ表示に変更）
+      showscale: true, // カラースケールを表示する
       colorbar: {
         title: {
-          text: '割合 (%)'
+          text: '割合 (%)' // カラーバーのタイトルを設定
         }
       }
     };
@@ -131,13 +139,13 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
 
   // ボタンの設定
   const updatemenus = [{
-    buttons: lineages.map((lineage, index) => ({
+    buttons: prefectures.map((prefecture, index) => ({
       method: 'update' as const,
       args: [
         {'visible': heatmapData.map((_, i) => i === index)},
-        {'title.text': `第${wave}波における${lineage}の割合`}
+        {'title.text': `第${wave}波における${prefecture}の系統別割合`}
       ],
-      label: lineage
+      label: prefecture
     })),
     direction: 'down' as const,
     showactive: true,
@@ -170,7 +178,7 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
     },
     yaxis: {
       title: {
-        text: '都道府県',
+        text: '系統',
         font: {
           color: '#000000',
           size: 14
@@ -208,9 +216,9 @@ export default async function HeatmapPage({ params }: { params: { wave: string }
         />
         {/* デバッグ: Prismaからデータが取得できているか表示 */}
         <div style={{ marginTop: '10px', fontSize: '12px', color: '#666' }}>
-          取得データ数: {data.length} | 日付数: {dates.length} | 系統数: {lineages.length} | 都道府県数: {topPrefectures.length}
+          取得データ数: {data.length} | 日付数: {dates.length} | 系統数: {lineages.length} | 都道府県数: {prefectures.length}
         </div>
       </div>
     </div>
   );
-} 
+}
